MEED-667 Move overview wallet to the common wallet folder
