+++ wallet-webapps-common/src/main/webapp/vue-app/wallet-overview/components/WalletOverviewRewardItem.vue
