+++ wallet-webapps/src/main/webapp/WEB-INF/portlet.xml
@@ -127,25 +127,6 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
     </portlet-info>
   </portlet>
   <portlet>
-    <portlet-name>WalletOverview</portlet-name>
-    <portlet-class>org.exoplatform.commons.api.portlet.GenericDispatchedViewPortlet</portlet-class>
-    <init-param>
-      <name>portlet-view-dispatched-file-path</name>
-      <value>/WEB-INF/jsp/walletOverview.jsp</value>
-    </init-param>
-    <init-param>
-      <description>List of profiles that can be used to enable this portlet</description>
-      <name>exo.profiles</name>
-      <value>wallet</value>
-    </init-param>
-    <supports>
-      <mime-type>text/html</mime-type>
-    </supports>
-    <portlet-info>
-      <title>Wallet overview</title>
-    </portlet-info>
-  </portlet>
-  <portlet>
     <portlet-name>WalletBalance</portlet-name>
     <portlet-class>org.exoplatform.commons.api.portlet.GenericDispatchedViewPortlet</portlet-class>
     <init-param>
